+++ src/pages/Dashboard/Sidebar/SidebarIcons.js
@@ -12,6 +12,22 @@ import { ReactComponent as Privacy } from "../../../static/icons/privacy.svg";
 import { ReactComponent as AccountDetails } from "../../../static/icons/mdi_account-details-outline.svg";
 import { ReactComponent as Terms } from "../../../static/icons/terms.svg";
 import { ReactComponent as Password } from "../../../static/icons/lock_icon.svg";
-import {BiBookmark as SavedJobs} from "react-icons/bi"
+import { BiBookmark as SavedJobs } from "react-icons/bi";
 
-export { ChangeProfilePicture, EditProfile, MyProfile, Dashboard, Applications, Analytics, Help, Settings, Messages, SavedJobs, Logout, Privacy, AccountDetails, Terms, Password};
+export {
+  ChangeProfilePicture,
+  EditProfile,
+  MyProfile,
+  Dashboard,
+  Applications,
+  Analytics,
+  Help,
+  Settings,
+  Messages,
+  SavedJobs,
+  Logout,
+  Privacy,
+  AccountDetails,
+  Terms,
+  Password,
+};
