+++ src/pages/Dashboard/TechTalent/RecommendationCard/index.module.scss
