+++ src/pages/Dashboard/TechTalent/setting/settings.scss
@@ -2,6 +2,7 @@
 .pass-header {
     border-bottom: 1px solid #000;
     max-width: 60vw;
+    padding: 10px;
     h2 {
         color: #004259;
         font-size: 24px;
@@ -10,35 +11,8 @@
     }
 }
 .password-section {
-    margin-top: 3.5%;
-    .dash-password {
-        width: 100%;
-        max-width: 780px;
-        label {
-            font-size: 18px;
-            font-weight: 500;
-            line-height: 24.59px;
-            color: #000;
-        }
-        .password-input {
-            position: relative;
-            margin: 0.5rem 0 1rem 3rem;
-            input {
-                padding: 0.8rem 0.7rem;
-                border: 1px solid #c9c9c9;
-                border-radius: 5px;
-                width: 100%;
-                align-items: center;
-                &::placeholder {
-                    color: #000;
-                }
-            }
-            .pass-toggle {
-                top: 52%;
-                left: 92%;
-            }
-        }
-    }
+    max-width: 60vw;
+    padding: 10px;
     .password-change {
         margin: 2rem 0;
         h3 {
