+++ src/pages/Dashboard/index.jsx
@@ -1,36 +1,56 @@
-import Sidebar from "../../components/DashboardComponents/Sidebar";
+import Sidebar from "./Sidebar";
 import s from "./index.module.scss";
 import pic from "../../static/images/Sarah.png";
-import { Outlet } from "react-router-dom";
+import { Outlet, useNavigate } from "react-router-dom";
 import { createContext } from "react";
 import axios from "axios";
 import { useState } from "react";
 import { useEffect } from "react";
-export const UserContext = createContext({});
+export const UserContext = createContext({
+  firstName: "",
+  lastName: "",
+  dateOfBirth: "",
+  email: "",
+});
 const Dashboard = () => {
-  const [user, setUser] = useState({});
+  const [user, setUser] = useState({
+    firstName: "",
+    lastName: "",
+    dateOfBirth: "",
+    email: "",
+  });
+
+  const [error, setErr] = useState("");
+
   const auth = window.localStorage.getItem("NXGJOBHUBLOGINKEYV1");
+  const navigate = useNavigate()
   useEffect(() => {
-    axios
+    auth ? axios
       .get("https://job-hub-591ace1cfc95.herokuapp.com/api/v1/auth/get-user", {
         headers: { authorization: auth },
       })
-      .then((res) => setUser(res.data));
-  }, [auth]);
-  return (
-    <UserContext.Provider value={user}>
-      <div className={s.Dashboard}>
-        <div className={s.Sidebad}>
-          <Sidebar profilePic={pic} />
-        </div>
+      .then((res) => setUser(res.data))
+      .catch((err) => setErr(err.message))
+      : navigate("/login")
+  }, [auth, navigate]);
 
-        {/* Mainpage or <Outlet/> */}
-        <div className={s.Sidemain}>
-          <Outlet />
+  return (auth ?
+    <>
+      <UserContext.Provider value={user}>
+        <div className={s.Dashboard}>
+          <div className={s.Sidebad}>
+            <Sidebar profilePic={pic} />
+          </div>
+
+          {/* Mainpage or <Outlet/> */}
+          <div className={s.Sidemain}>
+            <Outlet />
+          </div>
         </div>
-      </div>
-    </UserContext.Provider>
-  );
+      </UserContext.Provider>
+      <div className={s.error}>{error}</div>
+    </>
+  :null);
 };
 
 export default Dashboard;
